GUI refactoring. Moved GUI.java to gui package and created EventListeners.java to tidy up GUI.java a bit.
