+++ src/main/java/nl/tudelft/scrumbledore/gui/GUI.java
@@ -18,15 +18,12 @@ import javafx.scene.control.Label;
 import javafx.scene.control.RadioButton;
 import javafx.scene.control.Toggle;
 import javafx.scene.control.ToggleGroup;
-import javafx.scene.input.KeyEvent;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
-import javafx.stage.WindowEvent;
 import nl.tudelft.scrumbledore.Bubble;
-import nl.tudelft.scrumbledore.BubbleAction;
 import nl.tudelft.scrumbledore.Constants;
 import nl.tudelft.scrumbledore.Fruit;
 import nl.tudelft.scrumbledore.Game;
@@ -38,7 +35,6 @@ import nl.tudelft.scrumbledore.Player;
 import nl.tudelft.scrumbledore.PlayerAction;
 import nl.tudelft.scrumbledore.SpriteStore;
 import nl.tudelft.scrumbledore.StepTimer;
-import nl.tudelft.scrumbledore.Vector;
 
 /**
  * Launches the Scrumbledore GUI and performs all required handling actions that are related to the
@@ -48,8 +44,8 @@ import nl.tudelft.scrumbledore.Vector;
  * @author Jesse Tilro
  * @author Niels Warnars
  */
-@SuppressWarnings({ "checkstyle:methodlength", "PMD.TooManyMethods", "PMD.NPathComplexity", 
-  "PMD.CyclomaticComplexity", "PMD.ModifiedCyclomaticComplexity", "PMD.StdCyclomaticComplexity" })
+@SuppressWarnings({ "checkstyle:methodlength", "PMD.TooManyMethods", "PMD.NPathComplexity",
+    "PMD.CyclomaticComplexity", "PMD.ModifiedCyclomaticComplexity", "PMD.StdCyclomaticComplexity" })
 public class GUI extends Application {
   private Game game;
   private StepTimer timer;
@@ -95,9 +91,10 @@ public class GUI extends Application {
     setupGUI();
 
     // Add event listeners.
-    addKeyEventListeners(scene);
+    EventListeners el = new EventListeners(game, stage, scene);
+    el.init();
+
     addButtonEventListeners();
-    addWindowEventListeners(stage);
 
     renderStatic();
 
@@ -256,8 +253,9 @@ public class GUI extends Application {
       spr = "player-right";
     }
 
-    painter.drawImage(sprites.getImage(spr), game.getCurrentLevel().getPlayer().getPosition()
-        .getX(), game.getCurrentLevel().getPlayer().getPosition().getY());
+    painter.drawImage(sprites.getImage(spr),
+        game.getCurrentLevel().getPlayer().getPosition().getX(),
+        game.getCurrentLevel().getPlayer().getPosition().getY());
   }
 
   /**
@@ -299,8 +297,8 @@ public class GUI extends Application {
       if (current.getMovementDirection().equals(NPCAction.MoveLeft)) {
         spr = "enemy-mighta-left";
       }
-      painter.drawImage(sprites.getImage(spr), current.getPosition().getX(), current.getPosition()
-          .getY());
+      painter.drawImage(sprites.getImage(spr), current.getPosition().getX(),
+          current.getPosition().getY());
     }
   }
 
@@ -314,8 +312,8 @@ public class GUI extends Application {
     // Placing the platform elements within the level.
     for (Platform current : game.getCurrentLevel().getPlatforms()) {
       // Painting the current platform image at the desired x and y location given by the vector.
-      painter.drawImage(sprites.getImage("wall-1"), current.getPosition().getX(), current
-          .getPosition().getY());
+      painter.drawImage(sprites.getImage("wall-1"), current.getPosition().getX(),
+          current.getPosition().getY());
     }
   }
 
@@ -333,8 +331,8 @@ public class GUI extends Application {
     }
 
     for (Fruit current : fruits) {
-      painter.drawImage(sprites.getImage("fruit-banana"), current.getPosition().getX(), current
-          .getPosition().getY());
+      painter.drawImage(sprites.getImage("fruit-banana"), current.getPosition().getX(),
+          current.getPosition().getY());
     }
   }
 
@@ -462,91 +460,6 @@ public class GUI extends Application {
         Logger.log("--------------------GAME TERMINATED");
 
         System.exit(0);
-      }
-
-    });
-  }
-
-  /**
-   * Add key event listeners to a given scene to allow player controls.
-   * 
-   * @param scene
-   *          The scene the listeners should be added to.
-   */
-  private void addKeyEventListeners(Scene scene) {
-    // KeyPress Event handlers.
-    scene.setOnKeyPressed(new EventHandler<KeyEvent>() {
-
-      // Handling the key press.
-      public void handle(KeyEvent keyPressed) {
-        String keyPress = keyPressed.getCode().toString();
-        Player player = game.getCurrentLevel().getPlayer();
-        Vector bubblePos = player.getPosition().clone();
-        ArrayList<Bubble> bubbles = game.getCurrentLevel().getBubbles();
-
-        // Mapping the desired keys to the desired actions.
-        if (keyPress.equals("LEFT")) {
-          player.addAction(PlayerAction.MoveLeft);
-        } else if (keyPress.equals("RIGHT")) {
-          player.addAction(PlayerAction.MoveRight);
-        } else if (keyPress.equals("UP")) {
-          player.addAction(PlayerAction.Jump);
-        }
-
-        // Mapping the shooting action keys.
-        if (keyPress.equals("Z")) {
-          if (!player.isFiring()) {
-            Bubble newBubble = new Bubble(bubblePos, new Vector(Constants.BLOCKSIZE,
-                Constants.BLOCKSIZE));
-
-            bubbles.add(newBubble);
-            if (player.getLastMove() == PlayerAction.MoveLeft) {
-              if (Constants.LOGGING_WANTSHOOTING) {
-                // Sending the shooting information to the logger.
-                Logger.log("Player shot in the western direction.");
-              }
-
-              newBubble.addAction(BubbleAction.MoveLeft);
-            } else {
-              if (Constants.LOGGING_WANTSHOOTING) {
-                // Sending the shooting information to the logger.
-                Logger.log("Player shot in the eastern direction.");
-              }
-
-              newBubble.addAction(BubbleAction.MoveRight);
-            }
-          }
-
-          player.setFiring(true);
-        }
-
-        // Restarting the game if "R" is pressed or when the player is dead.
-        if (keyPress.equals("R") || !game.getCurrentLevel().getPlayer().isAlive()) {
-          game.restart();
-          renderStatic();
-        }
-      }
-
-    });
-
-    // KeyRelease Event handlers.
-    scene.setOnKeyReleased(new EventHandler<KeyEvent>() {
-
-      // Handling the key press.
-      public void handle(KeyEvent keyReleased) {
-        String keyRelease = keyReleased.getCode().toString();
-        Player player = game.getCurrentLevel().getPlayer();
-
-        // Mapping the desired keys to the desired actions.
-        if (keyRelease.equals("LEFT")) {
-          player.addAction(PlayerAction.MoveStop);
-        } else if (keyRelease.equals("RIGHT")) {
-          player.addAction(PlayerAction.MoveStop);
-        }
-
-        if (keyRelease.equals("Z")) {
-          player.setFiring(false);
-        }
       }
 
     });
@@ -702,25 +615,4 @@ public class GUI extends Application {
     settingsScene.getStylesheets().add(Constants.CSS_LOCATION);
     settingsStage.show();
   }
-
-  /**
-   * Add WindowEvent listener to exit the application when the window is closed.
-   * 
-   * @param stage
-   *          The Stage used by the rest of the GUI.
-   */
-  private void addWindowEventListeners(Stage stage) {
-    stage.setOnCloseRequest(new EventHandler<WindowEvent>() {
-      public void handle(WindowEvent event) {
-
-        // Logging the termination of the game.
-        Logger.log("--------------------GAME TERMINATED");
-
-        // Quitting the game.
-        System.exit(0);
-      }
-    });
-
-  }
-
 }
